Moved StringHash.GetStableHashCode to Utils.
