--- Assets/Mirror/Runtime/StringHash.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 733f020f9b76d453da841089579fd7a7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
