+++ Assets/Mirror/Runtime/UNetwork.cs
@@ -189,6 +189,19 @@ namespace Mirror
             return new float[]{u, v, w};
         }
 
+        // string.GetHashCode is not guaranteed to be the same on all machines, but
+        // we need one that is the same on all machines. simple and stupid:
+        public static int GetStableHashCode(this string text)
+        {
+            unchecked
+            {
+                int hash = 23;
+                foreach (char c in text)
+                    hash = hash * 31 + c;
+                return hash;
+            }
+        }
+
         // headless mode detection
         public static bool IsHeadless()
         {
